+++ Polinom/global.h
@@ -1,2 +1,2 @@
-extern int addOp;
-extern int mulOp;
+extern long long int addOp;
+extern long long int mulOp;
